requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
reportlab==4.2.2
passlib==1.7.4
face_recognition==1.3.0
scikit-learn==1.5.1
matplotlib==3.9.1
virtualenv==20.26.3
pydantic==2.8.2
starlette==0.37.2
cmake==3.30.0
numpy==1.23.5
keras==2.13.1
h5py==3.8.0
botcity-framework-core==1.2.1
mysql-connector
imap_tools
smtplib
pillow
os
